plug in API call to get component detail
